src/main/java/com/tteam/movieland/service/DefaultCountryService.java
package com.tteam.movieland.service;

import com.tteam.movieland.entity.Country;
import com.tteam.movieland.repository.CountryRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@RequiredArgsConstructor
public class DefaultCountryService implements CountryService{
    private final CountryRepository countryRepository;

    @Override
    public List<Country> getAll() {
        return countryRepository.findAll();
    }
}
